move order placement to APIRouter - still need to add additional api params
