+++ src/contexts/APIContext/types.ts
@@ -9,7 +9,6 @@ import {
   IGetAssetResult,
   ISearchAssetsResult,
   IGetOrdersResult,
-  IPlaceOrderResult,
 } from "types/NXTAPI";
 
 export interface ContextValues {
@@ -24,6 +23,5 @@ export interface ContextValues {
   getAsset?: (assetId: string) => Promise<false | IGetAssetResult>;
   getOrders?: (assetId: number) => Promise<false | IGetOrdersResult>;
   searchAssets?: (queryString: string) => Promise<false | ISearchAssetsResult>;
-  placeBidOrder?: (assetID: number, quantityQNT: string, priceNQT: string, secret: string) => Promise<false | IPlaceOrderResult>;
   handleFetchAccountIDFromRS?: (address: string) => Promise<string | undefined>;
 }
