+++ src/contexts/APIRouterContext/types.ts
@@ -1,4 +1,7 @@
+import { IPlaceOrderResult } from "types/NXTAPI";
+
 export interface ContextValues {
   sendJUP?: (toAddress: string, amount: string) => Promise<true | undefined>;
   sendAsset?: (toAddress: string, amount: string, assetId: string) => Promise<true | undefined>;
+  placeBidOrder?: (assetID: number, quantityQNT: string, priceNQT: string, secret: string) => Promise<false | IPlaceOrderResult>;
 }
